requirements.txt
openai==1.30.3
python-dotenv==0.19.2
# pandas==1.3.3
flask==3.0.3
weaviate-client==4.8.1
flask-cors==5.0.0
gunicorn==20.1.0